+++ src/main/kotlin/com/github/andrelmv/kotbox/wrap/wrappers/WrapLaunchAction.kt
@@ -1,0 +1,5 @@
+package com.github.andrelmv.kotbox.wrap.wrappers
+
+import com.github.andrelmv.kotbox.wrap.WrapWithCoroutineAction
+
+class WrapLaunchAction : WrapWithCoroutineAction("launch")