+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- DSL builder generator
+- Update shortcuts for wrap with coroutine builders
+
 ## [2.0.1] - 2026-04-04
 
 ### Added